mmu: direct-map all pages except for pages containing the kernel
This change uses a two-stage mapping approach borrowed from mini-os.
It entails mapping the code's expected virtual address as well as
its physical address to its physical location in the MMU.  This is
done to enable relocation because we don't know where Xen will load
us but we do need to be able to run at the location dictated by the
compiled kernel binary.

This preserves the behavior that all (non-kernel) pages are mapped
to the same underlying physical addresses.  It also changes Xenstore
setup to just use the underlying physical page for the xenstore
buffer rather than trying to create a custom mapping in the MMU
since I'm having trouble getting that working right now.