gic: get base addresses from Xen device tree
This change introduces a new dependency on libfdt/dtc.  To fetch the
dependency, run 'setup.sh' in the root of the repository.  The DTC
sources will be built into the FreeRTOS binary by the build system.

This change removes the hard-coded GIC base addresses from the source in
lieu of a runtime device tree query.